requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.3.0
djangorestframework==3.11.0
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-decouple==3.4
python-dotenv==0.17.0
python3-openid==3.2.0
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.3.0
urllib3==1.26.4
whitenoise==5.2.0
